+++ home/navigation/defaultNavigationOptions.tsx
@@ -21,7 +21,15 @@ export default (theme: ColorTheme): StackNavigationOptions => {
     headerTitleStyle: {
       fontWeight: Platform.OS === 'ios' ? '600' : '400',
       color: Colors[theme].text,
+      ...(FeatureFlags.ENABLE_2022_NAVIGATION_REDESIGN && {
+        fontFamily: 'Inter-SemiBold',
+        color: theme === 'dark' ? darkTheme.text.default : lightTheme.text.default,
+      }),
     },
+    ...(FeatureFlags.ENABLE_2022_NAVIGATION_REDESIGN && {
+      headerTintColor: theme === 'dark' ? darkTheme.link.default : lightTheme.link.default,
+    }),
+
     headerStyleInterpolator: HeaderStyleInterpolators.forUIKit,
   };
 };