+++ home/screens/HomeScreen/DevelopmentServersHeader.tsx
@@ -17,7 +17,8 @@ export function DevelopmentServersHeader({ onHelpPress }: DevelopmentServersHead
         <Heading
           color="secondary"
           size="small"
-          style={{ marginRight: spacing[2], fontWeight: Platform.OS === 'ios' ? '600' : 'bold' }}>
+          style={{ marginRight: spacing[2] }}
+          type="InterSemiBold">
           Development servers
         </Heading>
       </Row>