+++ packages/expo-dev-menu/ios/DevMenuViewController.swift
@@ -73,11 +73,15 @@ class DevMenuViewController: UIViewController {
   // MARK: private
 
   private func initialProps() -> [String: Any] {
+    let isSimulator = TARGET_IPHONE_SIMULATOR > 0
+    
     return [
       "showOnboardingView": manager.shouldShowOnboarding(),
       "appInfo": manager.getAppInfo(),
       "devSettings": manager.getDevSettings(),
+      "menuPreferences": DevMenuPreferences.serialize(),
       "uuid": UUID.init().uuidString,
+      "isDevice": !isSimulator,
     ]
   }