+++ packages/expo-dev-menu/ios/Modules/DevMenuInternalModule.swift
@@ -68,28 +68,7 @@ public class DevMenuInternalModule: NSObject, RCTBridgeModule {
 
   @objc
   func setOnboardingFinished(_ finished: Bool) {
-    DevMenuSettings.isOnboardingFinished = finished
-  }
-
-  @objc
-  func getSettingsAsync(_ resolve: RCTPromiseResolveBlock, reject: RCTPromiseRejectBlock) {
-    resolve(DevMenuSettings.serialize())
-  }
-
-  @objc
-  func setSettingsAsync(_ dict: [String: Any], resolve: RCTPromiseResolveBlock, reject: RCTPromiseRejectBlock) {
-    if let motionGestureEnabled = dict["motionGestureEnabled"] as? Bool {
-      DevMenuSettings.motionGestureEnabled = motionGestureEnabled
-    }
-    if let touchGestureEnabled = dict["touchGestureEnabled"] as? Bool {
-      DevMenuSettings.touchGestureEnabled = touchGestureEnabled
-    }
-    if let keyCommandsEnabled = dict["keyCommandsEnabled"] as? Bool {
-      DevMenuSettings.keyCommandsEnabled = keyCommandsEnabled
-    }
-    if let showsAtLaunch = dict["showsAtLaunch"] as? Bool {
-      DevMenuSettings.showsAtLaunch = showsAtLaunch
-    }
+    DevMenuPreferences.isOnboardingFinished = finished
   }
 
   @objc