+++ packages/expo-updates/android/src/main/java/expo/modules/updates/codesigning/CodeSigningConfiguration.kt
@@ -5,20 +5,31 @@ import android.util.Log
 import expo.modules.structuredheaders.BooleanItem
 import expo.modules.structuredheaders.Dictionary
 import expo.modules.structuredheaders.StringItem
+import expo.modules.updates.codesigning.CertificateChain.Companion.expoProjectInformation
 import java.security.Signature
 
 private const val TAG = "CodeSigning"
+
+enum class ValidationResult {
+  VALID,
+  INVALID,
+  SKIPPED
+}
+
+data class SignatureValidationResult(val validationResult: ValidationResult, val expoProjectInformation: ExpoProjectInformation?)
 
 /**
  * Contains all information about code signing.
  * @param embeddedCertificateString Implicitly trusted self-signed root certificate. May be the code signing certificate if only using embedded certificate (no chain).
  * @param codeSigningMetadata Metadata about the code signing certificate. May be the root certificate or a leaf certificate.
  * @param includeManifestResponseCertificateChain Should a certificate chain in the manifest response be evaluated when validating the signature. If true, ignores codeSigningMetadata when evaluating cert chain.
+ * @param allowUnsignedManifests Should unsigned manifest responses be allowed. Used during Expo Go transition when not all expo-cli versions know how to sign using the new method yet the new Expo Go clients have certificates.
  */
 class CodeSigningConfiguration(
   private val embeddedCertificateString: String,
   private val codeSigningMetadata: Map<String, String>?,
   private val includeManifestResponseCertificateChain: Boolean,
+  private val allowUnsignedManifests: Boolean
 ) {
   private val algorithmFromMetadata: CodeSigningAlgorithm by lazy {
     CodeSigningAlgorithm.parseFromString(
@@ -32,7 +43,24 @@ class CodeSigningConfiguration(
     codeSigningMetadata?.get(CODE_SIGNING_METADATA_KEY_ID_KEY) ?: CODE_SIGNING_METADATA_DEFAULT_KEY_ID
   }
 
-  fun validateSignature(info: SignatureHeaderInfo, bodyBytes: ByteArray, manifestResponseCertificateChain: String?): Boolean {
+  fun validateSignature(signature: String?, bodyBytes: ByteArray, manifestResponseCertificateChain: String?): SignatureValidationResult {
+    if (signature == null) {
+      if (!allowUnsignedManifests) {
+        throw Exception("No expo-signature header specified")
+      } else {
+        // no-op
+        return SignatureValidationResult(ValidationResult.SKIPPED, null)
+      }
+    }
+
+    return validateSignatureInternal(
+      SignatureHeaderInfo.parseSignatureHeader(signature),
+      bodyBytes,
+      manifestResponseCertificateChain,
+    )
+  }
+
+  fun validateSignatureInternal(info: SignatureHeaderInfo, bodyBytes: ByteArray, manifestResponseCertificateChain: String?): SignatureValidationResult {
     val certificateChain = if (includeManifestResponseCertificateChain) {
       CertificateChain(
         separateCertificateChain(manifestResponseCertificateChain ?: "") + embeddedCertificateString
@@ -62,10 +90,15 @@ class CodeSigningConfiguration(
     // specify what algorithm should be used in the chain case. One approach may be that in the case of
     // chains served alongside the manifest we fork the behavior to trust the `info.algorithm` while keeping
     // `metadata.algorithm` for the embedded case.
-    return Signature.getInstance("SHA256withRSA").apply {
+    val isValid = Signature.getInstance("SHA256withRSA").apply {
       initVerify(certificateChain.codeSigningCertificate.publicKey)
       update(bodyBytes)
     }.verify(Base64.decode(info.signature, Base64.DEFAULT))
+
+    return SignatureValidationResult(
+      if (isValid) ValidationResult.VALID else ValidationResult.INVALID,
+      certificateChain.codeSigningCertificate.expoProjectInformation()
+    )
   }
 
   fun getAcceptSignatureHeader(): String {